Share stats timeline presentation helpers
